+++ Patches/ItemDroneIndestructiblePatch.cs
@@ -1,5 +1,4 @@
 using HarmonyLib;
-using UnityEngine;
 
 // ReSharper disable InconsistentNaming
 namespace LegalizeNuclearBombs.Patches
@@ -11,16 +10,11 @@ namespace LegalizeNuclearBombs.Patches
         private static void UpdatePatch(ItemDroneIndestructible __instance)
         {
             var drain = LegalizeNuclearBombs.configIndestructibleDroneBatteryDrain.Value;
-            
-            if (drain <= 1f || __instance.itemEquippable.isEquipped || !__instance.itemDrone.magnetActive ||
-                !__instance.itemDrone.magnetTargetPhysGrabObject.GetComponent<NukeValuable>())
-                return;
-            
-            __instance.itemDrone.itemBattery.batteryLife -= drain * Time.deltaTime;
-            
-            // if (SemiFunc.PerSecond(.5f, __instance))
-            //     LegalizeNuclearBombs.Debug($"Attached to NukeValuable, accelerated battery drain by {drain}",
-            //         __instance);
+            __instance.itemDrone.itemBattery.batteryDrainRate =
+                drain <= 1f || __instance.itemEquippable.isEquipped || !__instance.itemDrone.magnetActive ||
+                !__instance.itemDrone.magnetTargetPhysGrabObject.GetComponent<ValuableWarhead>()
+                    ? __instance.itemDrone.batteryDrainRate
+                    : __instance.itemDrone.batteryDrainRate * drain;
         }
     }
 }